+++ src/main/resources/db/migration/V4__Create_category_table.sql
@@ -1,6 +1,6 @@
 create table category
 (
-    id bigint auto_increment comment '主键自增',
+    id int auto_increment comment '主键自增',
     category_name varchar(16) not null comment '分类名称',
     article_count int default 0 null comment '文章数',
     gmt_create bigint null comment '创建时间',
